MostrarMensajeDeConfirmacionYEspera

+++ 07-JOURNAL/src/modules/daybook/views/EntryView.vue
@@ -48,6 +48,7 @@
 import { defineAsyncComponent } from 'vue';
 import { mapGetters, mapActions } from 'vuex';  // se asocian con las propiedades computadas
 import getDayMonthYear from '../helpers/getDayMonthYear'
+import Swal from 'sweetalert2'
 
 export default {
     props: {
@@ -104,6 +105,12 @@ export default {
             this.entry = entry
         },
         async saveEntry() {
+
+            new Swal({
+                title: 'Espere por favor',
+                allowOutsideClick: false
+            })
+            Swal.showLoading()
             
             if( this.entry.id ){
                 // Actualizar
@@ -118,15 +125,35 @@ export default {
                 //recirectTo => entry, param: id
                 this.$router.push({ name: 'entry', params: { id } })
             }
+
+            Swal.fire('Guardado', 'Entrada registrada con éxito', 'success')
             
         },
         async onDeleteEntry() {
-            console.log('delete', this.entry );
+            
+            // const result = await Swal.fire({ // Se desetructura para que quede un para que quede un codigo más limpio 
+            const {isConfirmed} = await Swal.fire({
+                title: '¿Está seguro?',
+                text: 'Una vez borrado, no se puede recuperar',
+                showDenyButton: true,
+                confirmButtonText: 'Si, estoy seguro'
+            })
 
-            await this.deleteEntry( this.entry.id )
-            // Redireccionar al usuario fuera de aquí...
-            this.$router.push({ name: 'no-entry' })
-            // redireccionar al entry 
+            if ( isConfirmed ) {
+                new Swal({
+                    title: 'Espere por favor',
+                    allowOutsideClick: false
+                })
+                Swal.showLoading()
+                await this.deleteEntry( this.entry.id )
+                // Redireccionar al usuario fuera de aquí...
+                this.$router.push({ name: 'no-entry' })
+                // redireccionar al entry 
+
+                Swal.fire('Eliminado','','success')
+            }
+
+            
         }   
     },
